use combined regexp
